+++ packages/graph-framework/package.json
@@ -53,14 +53,16 @@
   },
   "peerDependencies": {
     "@automerge/automerge": "^2",
-    "@automerge/automerge-repo": "^1",
-    "@automerge/automerge-repo-react-hooks": "^1",
+    "@automerge/automerge-repo": "^2",
+    "@automerge/automerge-repo-react-hooks": "^2",
     "react": "^18"
   },
   "devDependencies": {
-    "@automerge/automerge": "^2.2.8",
-    "@automerge/automerge-repo": "^1.2.1",
-    "@automerge/automerge-repo-react-hooks": "^1.2.1",
+    "@automerge/automerge": "^v2.2.9-alpha.3",
+    "@automerge/automerge-repo": "^2.0.0-alpha.14",
+    "@automerge/automerge-repo-network-websocket": "^2.0.0-alpha.14",
+    "@automerge/automerge-repo-react-hooks": "^2.0.0-alpha.14",
+    "@automerge/automerge-repo-storage-indexeddb": "^2.0.0-alpha.14",
     "@testing-library/jest-dom": "^6.5.0",
     "@testing-library/react": "^16.0.1",
     "@types/react": "^18.3.7",
@@ -75,7 +77,9 @@
     "@graph-framework/schema": "workspace:*",
     "@graph-framework/space-events": "workspace:*",
     "@graph-framework/utils": "workspace:*",
+    "@noble/hashes": "^1.5.0",
     "@xstate/store": "^2.6.1",
+    "effect": "^3.10.19",
     "xstate": "^5.19.0"
   }
 }
